Refactor Query Response
