+++ app/client/src/pages/Editor/IDE/EditorPane/components/ActionToolbar.tsx
@@ -35,6 +35,7 @@ const ActionToolbar = (props: Props) => {
           size="sm"
           startIcon="settings-2-line"
         />
+        <Button isIconButton kind="tertiary" size="sm" startIcon="book-line" />
         <Button
           isIconButton
           kind="tertiary"
